requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
SQLAlchemy==2.0.21
alembic==1.12.0
email-validator==2.0.0.post2
python-multipart==0.0.6
psycopg2-binary==2.9.9